GPATH format changed.
1. Problem and the cause

Though the --file option of gtags(1) limits the range of project,
global(1) and htags(1) with the -o option ignore the range.
It is because GPATH has only path names of the source files in the project,
and global(1) and htags(1) with the -o (--other) option need to traverse
file system by themselves without refering GPATH.

2. Solution

To solve the problem, I have an idea to make GPATH have all path
names in the range. If it is, global(1) and htags(1) can refer GPATH
without traversing file system. As a result, the range defined by the
--file option of gtags(1) is kept.
This brins another merit that the operation of global(1) and htags(1)
with -o option quickens more than now.

3. Imprementation (new GPATH)

a. GPATH has all path names in the range of the project.

        [GPATH]
        key             data
        --------------------
        ./aaa.c\0       11\0
                |
                v
        key             data
        --------------------
        ./aaa.c\0       11\0
        ./README\0      12\0            <=== include README

b. Each record of GPATH has a flag value to distinguish source files
   and the rest.

        [GPATH]
        key             data
        --------------------
        ./aaa.c\0       11\0
                |
                v
        key             data
        --------------------
        ./aaa.c\0       11\0
        ./README\0      12\0o\0         <=== 'o' means other files.